+++ x/group/keeper/genesis.go
@@ -10,7 +10,6 @@ import (
 	"cosmossdk.io/x/group"
 
 	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // InitGenesis initializes the group module's genesis state.
@@ -18,29 +17,29 @@ func (k Keeper) InitGenesis(ctx context.Context, cdc codec.JSONCodec, data json.
 	var genesisState group.GenesisState
 	cdc.MustUnmarshalJSON(data, &genesisState)
 
-	sdkCtx := sdk.UnwrapSDKContext(ctx)
+	store := k.storeService.OpenKVStore(ctx)
 
-	if err := k.groupTable.Import(sdkCtx.KVStore(k.key), genesisState.Groups, genesisState.GroupSeq); err != nil {
+	if err := k.groupTable.Import(store, genesisState.Groups, genesisState.GroupSeq); err != nil {
 		panic(errors.Wrap(err, "groups"))
 	}
 
-	if err := k.groupMemberTable.Import(sdkCtx.KVStore(k.key), genesisState.GroupMembers, 0); err != nil {
+	if err := k.groupMemberTable.Import(store, genesisState.GroupMembers, 0); err != nil {
 		panic(errors.Wrap(err, "group members"))
 	}
 
-	if err := k.groupPolicyTable.Import(sdkCtx.KVStore(k.key), genesisState.GroupPolicies, 0); err != nil {
+	if err := k.groupPolicyTable.Import(store, genesisState.GroupPolicies, 0); err != nil {
 		panic(errors.Wrap(err, "group policies"))
 	}
 
-	if err := k.groupPolicySeq.InitVal(sdkCtx.KVStore(k.key), genesisState.GroupPolicySeq); err != nil {
+	if err := k.groupPolicySeq.InitVal(store, genesisState.GroupPolicySeq); err != nil {
 		panic(errors.Wrap(err, "group policy account seq"))
 	}
 
-	if err := k.proposalTable.Import(sdkCtx.KVStore(k.key), genesisState.Proposals, genesisState.ProposalSeq); err != nil {
+	if err := k.proposalTable.Import(store, genesisState.Proposals, genesisState.ProposalSeq); err != nil {
 		panic(errors.Wrap(err, "proposals"))
 	}
 
-	if err := k.voteTable.Import(sdkCtx.KVStore(k.key), genesisState.Votes, 0); err != nil {
+	if err := k.voteTable.Import(store, genesisState.Votes, 0); err != nil {
 		panic(errors.Wrap(err, "votes"))
 	}
 
@@ -53,9 +52,9 @@ func (k Keeper) ExportGenesis(ctx context.Context, _ codec.JSONCodec) *group.Gen
 
 	var groups []*group.GroupInfo
 
-	sdkCtx := sdk.UnwrapSDKContext(ctx)
+	store := k.storeService.OpenKVStore(ctx)
 
-	groupSeq, err := k.groupTable.Export(sdkCtx.KVStore(k.key), &groups)
+	groupSeq, err := k.groupTable.Export(store, &groups)
 	if err != nil {
 		panic(errors.Wrap(err, "groups"))
 	}
@@ -63,22 +62,22 @@ func (k Keeper) ExportGenesis(ctx context.Context, _ codec.JSONCodec) *group.Gen
 	genesisState.GroupSeq = groupSeq
 
 	var groupMembers []*group.GroupMember
-	_, err = k.groupMemberTable.Export(sdkCtx.KVStore(k.key), &groupMembers)
+	_, err = k.groupMemberTable.Export(store, &groupMembers)
 	if err != nil {
 		panic(errors.Wrap(err, "group members"))
 	}
 	genesisState.GroupMembers = groupMembers
 
 	var groupPolicies []*group.GroupPolicyInfo
-	_, err = k.groupPolicyTable.Export(sdkCtx.KVStore(k.key), &groupPolicies)
+	_, err = k.groupPolicyTable.Export(store, &groupPolicies)
 	if err != nil {
 		panic(errors.Wrap(err, "group policies"))
 	}
 	genesisState.GroupPolicies = groupPolicies
-	genesisState.GroupPolicySeq = k.groupPolicySeq.CurVal(sdkCtx.KVStore(k.key))
+	genesisState.GroupPolicySeq = k.groupPolicySeq.CurVal(store)
 
 	var proposals []*group.Proposal
-	proposalSeq, err := k.proposalTable.Export(sdkCtx.KVStore(k.key), &proposals)
+	proposalSeq, err := k.proposalTable.Export(store, &proposals)
 	if err != nil {
 		panic(errors.Wrap(err, "proposals"))
 	}
@@ -86,7 +85,7 @@ func (k Keeper) ExportGenesis(ctx context.Context, _ codec.JSONCodec) *group.Gen
 	genesisState.ProposalSeq = proposalSeq
 
 	var votes []*group.Vote
-	_, err = k.voteTable.Export(sdkCtx.KVStore(k.key), &votes)
+	_, err = k.voteTable.Export(store, &votes)
 	if err != nil {
 		panic(errors.Wrap(err, "votes"))
 	}